+++ CHANGELOG.md
@@ -13,6 +13,7 @@ tech changes will usually be stripped from release notes for the public
 ### Added
 
 -   During panning and zooming a lower resolution render is now used if slower render times are detected
+-   [server] Added base server stats overview to admin panel
 
 ### Changed